+++ src/vivarium/framework/configuration.py
@@ -1,8 +1,19 @@
+"""
+=======================
+Configuration Utilities
+=======================
+
+A set of functions for turning model specification files into programmatic
+representations of :term:`model specifications <Model Specification>` and
+:term:`configurations <Configuration>`.
+
+"""
+
 import os.path
 
 import yaml
 
-from vivarium import VivariumError
+from vivarium.exceptions import VivariumError
 from vivarium.config_tree import ConfigTree
 
 from .plugins import DEFAULT_PLUGINS
@@ -32,7 +43,7 @@ def validate_model_specification_file(file_path: str) -> str:
     if extension not in ['yaml', 'yml']:
         raise ConfigurationError(f'Model specification files must be in a yaml format. You provided {extension}')
     # Attempt to load
-    yaml.load(file_path)
+    yaml.full_load(file_path)
     return file_path