+++ src/vivarium/framework/metrics.py
@@ -1,4 +1,14 @@
-"""A simple component to allow users to store outputs in the simulation."""
+"""
+==============
+Output Metrics
+==============
+
+Currently, ``vivarium`` uses the :ref:`values pipeline system <values_concept>`
+to produce the results, or output metrics, from a simulation.  The metrics
+The component here is a normal ``vivarium`` component whose only purpose is
+to provide an empty :class:`dict` as the source of the *"Metrics"* pipeline.
+It is included by default in all simulations.
+"""
 
 
 class Metrics: