+++ src/vivarium/framework/randomness.py
@@ -1,4 +1,9 @@
-"""This module contains classes and functions supporting common random numbers.
+"""
+==============================
+Random Numbers in ``vivarium``
+==============================
+
+This module contains classes and functions supporting common random numbers.
 
 Vivarium has some peculiar needs around randomness. We need to be totally consistent
 between branches in a comparison. For example, if a simulant gets hit by a truck
@@ -25,6 +30,11 @@ RESIDUAL_CHOICE : object
     ----
     Currently this object is only used in the `choice` function of this
     module.
+
+
+For mor information, see the Common Random Numbers
+:ref:`concept note <crn_concept>`.
+
 """
 from typing import Union, List, Tuple, Callable, Any
 import hashlib
@@ -33,7 +43,7 @@ import datetime
 import numpy as np
 import pandas as pd
 
-from vivarium import VivariumError
+from vivarium.exceptions import VivariumError
 from .utilities import rate_to_probability
 
 Array = Union[List, Tuple, np.ndarray, pd.Series]
@@ -182,7 +192,7 @@ class IndexMap:
         return 'IndexMap({})'.format("\n         ".join(repr(self._map).split("\n")))
 
 
-def random(key: str, index: Index, index_map: IndexMap=None) -> pd.Series:
+def random(key: str, index: Index, index_map: IndexMap = None) -> pd.Series:
     """Produces an indexed `pandas.Series` of uniformly distributed random numbers.
 
     The index passed in typically corresponds to a subset of rows in a
@@ -243,7 +253,7 @@ def get_hash(key: str) -> int:
     return int(hashlib.sha1(key.encode('utf8')).hexdigest(), 16) % 4294967295
 
 
-def choice(key: str, index: Index, choices: Array, p: Array=None, index_map: IndexMap=None) -> pd.Series:
+def choice(key: str, index: Index, choices: Array, p: Array = None, index_map: IndexMap = None) -> pd.Series:
     """Decides between a weighted or unweighted set of choices.
 
     Given a a set of choices with or without corresponding weights,
@@ -337,7 +347,7 @@ def _set_residual_probability(p: np.ndarray) -> np.ndarray:
 
 
 def filter_for_probability(key: str, population: Union[pd.DataFrame, pd.Series, Index],
-                           probability: Array, index_map: IndexMap=None) -> Union[pd.DataFrame, pd.Series, Index]:
+                           probability: Array, index_map: IndexMap = None) -> Union[pd.DataFrame, pd.Series, Index]:
     """Decide an event outcome for each individual in a population from probabilities.
 
     Given a population or its index and an array of associated probabilities
@@ -409,8 +419,8 @@ class RandomnessStream:
     `engine.Builder`
     """
     def __init__(self, key: str, clock: Callable, seed: Any,
-                 index_map: IndexMap=None, manager: 'RandomnessManager'=None,
-                 for_initialization: bool=False):
+                 index_map: IndexMap = None, manager: 'RandomnessManager' = None,
+                 for_initialization: bool = False):
         self.key = key
         self.clock = clock
         self.seed = seed
@@ -438,7 +448,7 @@ class RandomnessStream:
         else:
             return RandomnessStream(self.key, self.clock, self.seed, self.index_map)
 
-    def _key(self, additional_key: Any=None) -> str:
+    def _key(self, additional_key: Any = None) -> str:
         """Construct a hashable key from this object's state.
 
         Parameters
@@ -453,7 +463,7 @@ class RandomnessStream:
         """
         return '_'.join([self.key, str(self.clock()), str(additional_key), str(self.seed)])
 
-    def get_draw(self, index: Index, additional_key: Any=None) -> pd.Series:
+    def get_draw(self, index: Index, additional_key: Any = None) -> pd.Series:
         """Get an indexed sequence of floats pulled from a uniform distribution over [0.0, 1.0)
 
         Parameters
@@ -477,7 +487,7 @@ class RandomnessStream:
 
         return draw
 
-    def get_seed(self, additional_key: Any=None) -> int:
+    def get_seed(self, additional_key: Any = None) -> int:
         """Get a randomly generated seed for use with external randomness tools.
 
         Parameters
@@ -494,7 +504,7 @@ class RandomnessStream:
         return get_hash(self._key(additional_key))
 
     def filter_for_rate(self, population: Union[pd.DataFrame, pd.Series, Index],
-                        rate: Array, additional_key: Any=None) -> Index:
+                        rate: Array, additional_key: Any = None) -> Index:
         """Decide an event outcome for each individual in a population from rates.
 
         Given a population or its index and an array of associated rates for
@@ -528,7 +538,7 @@ class RandomnessStream:
         return self.filter_for_probability(population, rate_to_probability(rate), additional_key)
 
     def filter_for_probability(self, population: Union[pd.DataFrame, pd.Series, Index],
-                               probability: Array, additional_key: Any=None) -> Index:
+                               probability: Array, additional_key: Any = None) -> Index:
         """Decide an event outcome for each individual in a population from probabilities.
 
         Given a population or its index and an array of associated probabilities
@@ -555,7 +565,7 @@ class RandomnessStream:
         """
         return filter_for_probability(self._key(additional_key), population, probability, self.index_map)
 
-    def choice(self, index: Index, choices: Array, p: Array=None, additional_key: Any=None) -> pd.Series:
+    def choice(self, index: Index, choices: Array, p: Array = None, additional_key: Any = None) -> pd.Series:
         """Decides between a weighted or unweighted set of choices.
 
         Given a a set of choices with or without corresponding weights,
@@ -615,7 +625,7 @@ class RandomnessManager:
         self._clock = None
         self._key_columns = None
         self._key_mapping = IndexMap()
-        self._decision_points = set()
+        self._decision_points = dict()
 
     def setup(self, builder):
         self._seed = str(builder.configuration.randomness.random_seed)
@@ -627,7 +637,7 @@ class RandomnessManager:
         pop_size = builder.configuration.population.population_size
         self._key_mapping.map_size = max(map_size, 10*pop_size)
 
-    def get_randomness_stream(self, decision_point: str, for_initialization: bool=False) -> RandomnessStream:
+    def get_randomness_stream(self, decision_point: str, for_initialization: bool = False) -> RandomnessStream:
         """Provides a new source of random numbers for the given decision point.
 
         Parameters
@@ -651,9 +661,10 @@ class RandomnessManager:
         if decision_point in self._decision_points:
             raise RandomnessError(f"Two separate places are attempting to create "
                                   f"the same randomness stream for {decision_point}")
-        self._decision_points.add(decision_point)
-        return RandomnessStream(key=decision_point, clock=self._clock, seed=self._seed,
-                                index_map=self._key_mapping, manager=self, for_initialization=for_initialization)
+        stream = RandomnessStream(key=decision_point, clock=self._clock, seed=self._seed,
+                                  index_map=self._key_mapping, manager=self, for_initialization=for_initialization)
+        self._decision_points[decision_point] = stream
+        return stream
 
     def register_simulants(self, simulants: pd.DataFrame):
         """Adds new simulants to the randomness mapping.
@@ -675,7 +686,7 @@ class RandomnessManager:
 
     def __repr__(self) -> str:
         return f"RandomnessManager(seed={self._seed}, key_columns={self._key_columns}, " \
-               f"decision_points={self._decision_points})"
+               f"decision_points={self._decision_points.keys()})"
 
 
 class RandomnessInterface: