+++ src/vivarium/framework/time.py
@@ -1,4 +1,15 @@
-"""Vivarium time manager."""
+"""
+====================
+The Simulation Clock
+====================
+
+The components here provide implementations of different kinds of simulation
+clocks for use in ``vivarium``.
+
+For more information about time in the simulation, see the associated
+:ref:`concept note <time_concept>`.
+
+"""
 from typing import Union, Callable
 from numbers import Number
 from datetime import datetime, timedelta
@@ -6,8 +17,8 @@ from datetime import datetime, timedelta
 import pandas as pd
 
 
-_Time = Union[datetime, Number]
-_Timedelta = Union[timedelta, Number]
+Time = Union[datetime, Number]
+Timedelta = Union[timedelta, Number]
 
 
 class SimulationClock:
@@ -19,19 +30,19 @@ class SimulationClock:
         self._step_size = None
 
     @property
-    def time(self) -> _Time:
+    def time(self) -> Time:
         """The current simulation time."""
         assert self._time is not None, 'No start time provided'
         return self._time
 
     @property
-    def stop_time(self) -> _Time:
+    def stop_time(self) -> Time:
         """The time at which the simulation will stop."""
         assert self._stop_time is not None, 'No stop time provided'
         return self._stop_time
 
     @property
-    def step_size(self) -> _Timedelta:
+    def step_size(self) -> Timedelta:
         """The size of the next time step."""
         assert self._step_size is not None, 'No step size provided'
         return self._step_size