+++ src/vivarium/framework/utilities.py
@@ -1,3 +1,11 @@
+"""
+===========================
+Framework Utility Functions
+===========================
+
+Collection of utility functions shared by the ``vivarium`` framework.
+
+"""
 from importlib import import_module
 from typing import Callable
 
@@ -13,6 +21,10 @@ def to_yearly(value, time_step):
 
 
 def rate_to_probability(rate):
+    # encountered underflow from rate > 30k
+    # for rates greater than 250, exp(-rate) evaluates to 1e-109
+    # beware machine-specific floating point issues
+    rate[rate > 250] = 250.0
     return 1-np.exp(-rate)